+++ src/config.ts
@@ -26,8 +26,7 @@ class BaseConfiguration {
    * Device Pixel Ratio
    * @see https://developer.apple.com/library/safari/documentation/AudioVideo/Conceptual/HTML-canvas-guide/SettingUptheCanvas/SettingUptheCanvas.html
    */
-  devicePixelRatio =
-    typeof window === 'undefined' ? 1 : window.devicePixelRatio; // eslint-disable-line no-restricted-globals
+  devicePixelRatio?: number;
 
   /**
    * Pixel limit for cache canvases. 1Mpx , 4Mpx should be fine.