+++ src/env/index.ts
@@ -50,7 +50,16 @@ export const getFabricWindow = (): TFabricWindow => getEnv().window;
 /**
  * @returns the config value if defined, fallbacks to the environment value
  */
-export const getDevicePixelRatio = () =>
-  Math.max(config.devicePixelRatio ?? getFabricWindow().devicePixelRatio, 1);
+export const getDevicePixelRatio = () => {
+  if (config.devicePixelRatio !== undefined) {
+    return Math.max(config.devicePixelRatio, 1);
+  }
+
+  try {
+    return Math.max(getFabricWindow().devicePixelRatio ?? 1, 1);
+  } catch {
+    return 1;
+  }
+};
 
 export type * from './types';